% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ratios_distances.R
\name{calc_ratio_dist}
\alias{calc_ratio_dist}
\title{calc_ratio_distance}
\usage{
calc_ratio_dist(ratio_dat)
}
\arguments{
\item{ratio_dat}{The output of calc_ratios, which includes the four ratios of interest for both sources
and samples.}
}
\value{
A list of three data frames, which includes 1) distances for each
double ratio plot, source, and sample combination, along with the ranking of each source by sample (1 = closest) for each
double ratio plot ("raw"), 2) distances summarized by source, which contains the mean
distance, rank, and the number of top ranks calculated for each source and double ratio plot combination ("source"),
3) distances summarized by sample, which reports the top-ranked source for each sample and double
ratio combination ("sample"). Abbreviations for the ratios include: Anth_AnthPhen or AAP (ratio of anthracene
to anthracene plus phenanthrene), Flua_FluaPyr or FFP (ratio of fluoranthene to fluoranthene plus pyrene),
Baa_BaaCh or BBC (ratio of benzo(a)anthracene to benzo(a)anthracene plus chrysene), Indpy_IndpyBghip or IIB (ratio
of indeno(1,2,3-cd)pyrene to indeno(1,2,3-cd)pyrene plus benzo(g,h,i)perylene). Each double ratio comparison
is distinguished by the shorter of the abbreviations separated by an underscore (e.g., FFP_AAP is one double ratio
comparison).
}
\description{
Calculates the Euclidean distances between samples and sources from
the double ratio plots.
}
