% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_thresholds.R
\name{calc_tox_thresholds}
\alias{calc_tox_thresholds}
\title{calc_tox_thresholds}
\usage{
calc_tox_thresholds(compound_info, sample_column, conc_column, compound_column,
  conc_unit = "ppb")
}
\arguments{
\item{compound_info}{The output dataframe from `get_compound_info`, which contains sample
concentrations as well as compound-specific information, including whether the compound is one of
16 EPA priority compounds and compound-specific toxicity.}

\item{sample_column}{string, column that contains unique sample identifier}

\item{conc_column}{string, column that contains sample concentrations}

\item{compound_column}{string, column that contains compound names. For this function to work properly,
there must be a compound named "TOC" for total organic carbon (for TOC normalization in ESBTU calculation),
with units of percent TOC.}

\item{conc_unit}{string, the units of PAH concentrations,
either "ppb" (ug/kg) or "ppm" (mg/kg).}
}
\value{
a list of two dataframes. 'results_bysite' reports sample-specific thresholds, while
'results_summary' reports summary statistics of thresholds across all samples.
}
\description{
Summarizes PAH concentrations relative to toxicity thresholds including
1) the concentration above which adverse outcomes to aquatic biota are expected
(probable effect concentration or PEC), 2) the concentration below which adverse
outcomes to aquatic biota are unlikely (threshold effect concentration or TEC),
and 3) a threshold that accounts for the bioavailability of PAHs based on the
organic carbon content of the sediment (sum equilibrium-partitioning sediment
benchmark toxicity unit or ESBTU).
}
