% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_info.R
\name{get_compound_info}
\alias{get_compound_info}
\title{get_compound_info}
\usage{
get_compound_info(pah_dat, merge_type = "name", merge_col = "parameter_nm")
}
\arguments{
\item{pah_dat}{data frame of user PAH data, either retrieved using the get_pah data, or
the user's own PAH concentration data. Data frame must include a PAH identifier column, which can be
either the USGS parameter code, casrn, or compound name. Not all compounds have casrn numbers. This function
requires that pah_dat be in long format; that is, there is a parameter column and site column, where site
or sample id is repeated for each compound.}

\item{merge_type}{whether the merge should being done by USGS parameter codes ('pcode'), Chemical Abstracts
Service Registry Number ('casrn'), or compound name ('name'). Note that alkylated PAHs
(e.g., C2-alkylated naphthalene) do not have
casrn numbers, and metadata on these compounds will not be merged if merge_type = "casrn".}

\item{merge_col}{column name that contains the PAH identifier by which to merge}
}
\value{
A data frame with all of the columns from the user-supplied pah_dat, as well as columns from the
built in data frame pah_compounds. These included...
}
\description{
Merges user PAH data with compound-specific metadata, including whether the compound is: one of
the 16 EPA priority compounds, low or high molecular weight, a parent or alkylated PAH, and whether
the compound is in the source profiles.
}
