% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pah.R
\name{getPAH}
\alias{getPAH}
\title{get_pah}
\usage{
getPAH(sites, pcodes = pah_pcodes, startDate, endDate)
}
\arguments{
\item{pcodes}{vector of pcodes to call from NWIS, default set to 84 pcodes that specify PAH-related pcodes}

\item{startDate}{start date from which to pull data, in YYY-MM-DD format}

\item{site}{vector of character strings representing USGS site IDs}
}
\description{
Retrieve PAH data from WQP. If you already have your PAH data in-hand, there is no
need to run this function.
}
