% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah.R
\docType{data}
\name{pah_compounds}
\alias{pah_compounds}
\title{PAH compound metadata}
\format{A data frame with 84 observations of 18 variables:
\describe{
  \item{pcode}{USGS parameter code}
  \item{schedule}{USGS lab schedule}
  \item{Parameter}{Compound name}
  \item{Abbreviation}{Compound short name}
  \item{EPApriority16}{logical, is this compound one of the 16 EPA priority compounds?}
  }}
\usage{
pah_compounds
}
\description{
A table of PAH compound names, abbreviations, pcodes, and supporting information such as
molecular weights and whether the compound is a parent or alkylated compound.
}
\examples{
head(pah_compounds$Parameter)
}
\keyword{datasets}
