% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah_mw_parent.R
\name{pah_mw_parent}
\alias{pah_mw_parent}
\title{pah_mw_parent}
\usage{
pah_mw_parent(compound_info, sample_column, conc_column,
  statistic = "average", plot = TRUE)
}
\arguments{
\item{compound_info}{The output dataframe from `get_compound_info`, which contains sample concentrations
as well as compound-specific information, including low versus high molecular weight
(column "molwt_highlow") and parent versus alkylated (column "parentAlkyl").}

\item{sample_column}{string, column that contains unique sample identifier}

\item{conc_column}{string, column that contains sample concentrations}

\item{statistic}{string, either "sum" or "average". For each sample, compounds in each category
(low vs. high, parent vs. alkyl) are summed. Because different analyses may measure a different number
of chemicals or focus on different types of chemicals, these numbers are hard to interpret both within
and across studies. The raw sums can be reported ("sum") along with the number of compounds in each category.
Alternatively, the average concentration of compounds in each category ("average") as a way to standardize
the number of compounds across categories. If plot = FALSE, both are calculated and exported. If plot = TRUE,
only the statistic of choice is plotted.}

\item{plot}{logical, whether the statistic should be exported as a dataframe (FALSE) or
plotted as a boxplot (TRUE).}
}
\description{
Summarizes compound characteristics, including low versus high molecular weight and
parent versus alkylated, as one line of evidence for potential sources. Samples that
contain high molecular weight and parent compounds indicate a pyrogenic rather than petrogenic
PAH source.
}
