% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah_pca.R
\name{pah_pca}
\alias{pah_pca}
\title{pah_pca}
\usage{
pah_pca(profiles, perc_cutoff = 10, sample_column)
}
\arguments{
\item{profiles}{The output of pah_profiles, which contains PAH profile data for both source and samples.
The reported values are proportional concentrations for each compound.}

\item{perc_cutoff}{The threshold for keeping PCA components in subsequent analyses; all
individual components that explain greater than or equal to this cutoff of the total
variance in the dataset will be kept in the analysis.}

\item{sample_column}{column name of unique sample identifier}
}
\value{
A list of three data frames. "pca_dat" contains all PCA components that met the perc_cutoff,
as well as a sample_id column and type which identifies whether the values correspond to a sample
or source profile. "pca_summary" prints the standard deviation, proportion of variance, and cumulative
proportion of variance explained for all components of the PCA. "pca_distance" is a long data frame of all
source-sample compindations, and the euclidean distance between the source and sample in the PCA
components space.
}
\description{
Conducts principle components analysis on the source and sample PAH profiles. This reduces the profiles
to manageable dimensions. All important PCA axes are then used to calculate the Euclidean distance between
source and samples to determine likely sources.
}
