% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah.R
\docType{data}
\name{pah_pcodes}
\alias{pah_pcodes}
\title{Default PAH pcodes}
\format{An object of class \code{character} of length 84.}
\usage{
pah_pcodes
}
\description{
A vector of pcodes related to PAH analysis identified in Baldwin et al. (2017).
}
\examples{
length(pah_pcodes)
}
\keyword{datasets}
