% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pah.R
\name{plot_pah}
\alias{plot_pah}
\title{plot_pah}
\usage{
plot_pah(pah_dat, conc_column = "Value", sample_id_column = "Sample",
  compound_column = "Parameter", compound_plot = "Total PAH",
  color_column = NA, group_column = NA, conc_units = "ppb")
}
\arguments{
\item{pah_dat}{dataframe with PAH concentrations. It is recommended that entire samples with censored observations
(e.g., below detection limit) are removed from this data frame prior to calculating the difference between source and samples.
This function assumes these samples have already been removed.}

\item{conc_column}{column that contains PAH concentrations}

\item{sample_id_column}{column name that contains unique sample id}

\item{compound_column}{column name that contains PAH compound names. This can also include other chemicals, or sums of chemicals.}

\item{compound_plot}{a vector of strings identifying which compounds from compound_column to include in the plot. If more than one compound is given, the plot will be faceted by compound.}

\item{color_column}{a column with group variable by which to color code bars}

\item{group_column}{a column by which to group and order the bars.}
}
\description{
Create a bar plot of PAH concentrations. The function allows you to group the data by one
or more variables, and decide which grouping variables will determine order or color.
}
\examples{

}
