% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{plot_profiles}
\usage{
plot_profiles(profile_dat, plot_type = "boxplot", sources_plot = NA,
  samples_plot = "all", sample_column = "sample_id",
  source_abbreviation = FALSE)
}
\arguments{
\item{profile_dat}{The output of `pah_profiler` which includes the proportional concentration
profiles from the samples and sources, as well the sum of the distance metric for each
sample/source combination. These are combined in a list.}

\item{plot_type}{The desired plot type. This can either be a boxplot ('boxplot') of all distance measures
by source, or various profile ('profile') plots.}

\item{sources_plot}{If `plot_type` = `profile`, include a vector of abbreviated sources that you want to include
in a profile plot to compare to samples. Set equal to "all" to include paneled plots of all source profiles.}

\item{samples_plot}{If `plot_type` = `profile`, a unique sample ID to use to plot against source profiles.
Can either be a single unique ID or "all" to indicate taking the mean and standard deviation of all samples.}

\item{sample_column}{Column that contains unique sample IDs.}

\item{source_abbreviation}{logical, whether source abbreviations should be used when plot_type = 'distance_boxplot'.}
}
\description{
Creates common figures from the calculated distance between sample and source profiles.
Possible plots include a boxplot of the sum of the distance metric across all sources and
proportional concentration profile figures comparing an individual sample to a source.
}
