% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratios.R
\name{plot_ratios}
\alias{plot_ratios}
\title{plot_ratios}
\usage{
plot_ratios(ratio_dat, source_drop = "COA2")
}
\arguments{
\item{ratio_dat}{A dataframe which is the output from `calc_ratios`}

\item{source_drop}{A vector of source abbreviations to drop from plotting. This allows the user
to drop source outliers, or sources that are not near any samples.}
}
\value{
A 1x3 plot of relevant diagnostic double ratio plots. Note that this function defaults to drop
the source "COA2" because it is an outlier relative to other sources. Use 'source_drop = NA' to include
all sources.
}
\description{
This function uses takes the output of `calc_ratios` and creates double-ratio plots. This allows
comparison between samples and source ratios, and calculate which source ratio is closest (in Euclidean
distance) to the sample.
}
