% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah.R
\docType{data}
\name{source_ratios}
\alias{source_ratios}
\title{PAH source ratios}
\format{A data frame with 36 observations of 9 variables:
\describe{
  \item{sourceCategory}{Type of material sampled.}
  \item{sourceNo}{Unique identifier for the source material.}
  \item{sourceID}{Source descriptor.}
  \item{abbrev}{Source abbreviation (for plotting).}
  \item{Reference}{Published paper which contains the source ratios.}
  \item{Anth_AnthPhen}{The ratio of anthracene to anthracene plus phenanthrene.}
  \item{Flua_FluaPyr}{The ratio of fluoranthene to fluoranthene plus pyrene.}
  \item{Baa_BaaCh}{The ratio of benzo(a)anthracene to benzo(a)anthracene plus chrysene.}
  \item{Indpy_IndpyBghip}{The ratio of indeno(1,2,3-cd)pyrene to indeno(1,2,3-cd)pyrene plus benzo(g,h,i)perylene.}
  }}
\usage{
source_ratios
}
\description{
A table of various PAH sources and associated PAH ratios. These source ratios come from various sources,
and the citation is included in the table. These ratios can be used to calculate distance in a
double ratio plot.
}
\examples{
head(source_ratios)
}
\keyword{datasets}
