% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah.R
\docType{data}
\name{sources}
\alias{sources}
\title{PAH source names}
\format{A data frame with 52 observations of 4 variables:
\describe{
  \item{source_subcategory}{Coarse categories that describe the origin of the source (e.g., coal combustion)}
  \item{source_name}{Long name of source}
  \item{source_short_name}{Shortened name of source}
  \item{source_abbrev}{An abbreviated source name for plotting purposes, as well as merging with other sources tables.}
  }}
\usage{
sources
}
\description{
A table of various PAH sources, shortened and abbreviated names, as well as source categories
to clean up plotting or to group compounds.
}
\examples{
head(sources)
}
\keyword{datasets}
