% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ratios_distances.R
\name{calc_ratio_dist}
\alias{calc_ratio_dist}
\title{calc_ratio_distance}
\usage{
calc_ratio_dist(ratio_dat, summary_type = "sample")
}
\arguments{
\item{ratio_dat}{The output of calc_ratios, which includes the four ratios of interest for both sources
and samples.}

\item{summary_type}{Whether the distances should be exported as the raw distances and ranks for each
double ratio plot, source, and sample combination ("raw"),
or summarized by "sample" or by "source". If the data are summarized by source,
the generated table contains a source in each row, and the mean distance and rank between that source
and all samples. This will help discern which sources are closest to the samples overall (or on average). If the distances are summarized
If the data are summarized by sample, the generated table contains a sample in each row, and the top ranked
source for each double ratio comparison. This will help discern the likely source for each sample.}
}
\value{
A data frame that contains a column `sample_id` which contains the original sample identifiers,
as well as the unique identifiers of each source contained in `source_profiles`. The source and sample
ratios are defined in the `sample_type` column which is defined as either `source` or `sample`. Additionally,
four columns of ratios are included: Anth_AnthPhen (ratio of anthracene to anthracene plus phenanthrene),
Flua_FluaPyr (ratio of fluoranthene to fluoranthene plus pyrene), Baa_BaaCh (ratio of benzo(a)anthracene
to benzo(a)anthracene plus chrysene), Indpy_IndpyBghip (ratio of indeno(1,2,3-cd)pyrene to
indeno(1,2,3-cd)pyrene plus benzo(g,h,i)perylene).
}
\description{
Calculates the Euclidean distances between samples and sources from
the double ratio plots.
}
