% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah_mw_parent.R
\name{pah_mw_parent}
\alias{pah_mw_parent}
\title{pah_mw_parent}
\usage{
pah_mw_parent(compound_info, sample_column, conc_column,
  statistic = "average", plot = TRUE)
}
\arguments{
\item{compound_info}{The output dataframe from `get_compound_info`, which contains sample concentrations
as well as compound-specific information, including low versus high molecular weight
(column "molwt_highlow") and parent versus alkylated (column "parentAlkyl").}

\item{sample_column}{string, column that contains unique sample identifier}

\item{conc_column}{string, column that contains sample concentrations}

\item{statistic}{string, either "sum" or "average". If statistic = "sum", compounds are summed by sample and
category (low vs. high, parent vs. alkyl for each sample). Because different analyses may measure a different number
of chemicals or focus on different types of chemicals, these numbers are hard to interpret both within
and across studies. The raw sums are reported ("sum") along with the number of compounds in each category.
Alternatively, the average concentration of compounds in each category ("average") is a way to standardize
the number of compounds across categories. If plot = FALSE, both statistics  are calculated and exported
in the first table in the list ("all_dat"), but only the chosen statistic is summarized and exported in
the second table that summarizes HMW:LMW and parent:alkyl for each sample ("summarized_dat"). If plot = TRUE,
only the statistic of choice is plotted.}

\item{plot}{logical, whether the data should be exported as two dataframes within a list (FALSE) or
plotted as a boxplot (TRUE).}
}
\description{
Summarizes compound characteristics, including low versus high molecular weight and
parent versus alkylated, as one line of evidence for potential sources. Samples that
contain high molecular weight and parent compounds indicate a pyrogenic rather than petrogenic
PAH source.
}
