% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah_pca.R
\name{pah_pca}
\alias{pah_pca}
\title{pah_pca}
\usage{
pah_pca(profiles, perc_cutoff = 10, include_creosote = T)
}
\arguments{
\item{profiles}{The output of pah_profiles, which contains PAH profile data for both source and samples.
The reported values are proportional concentrations for each compound.}

\item{perc_cutoff}{The threshold for keeping PCA components in subsequent analyses; all
individual components that explain greater than or equal to this cutoff of the total
variance in the dataset will be kept in the analysis.}

\item{include_creosote}{Logical, whether to include the source profiles for creosote (n = 2). The source profiles
for creosote only include 11 compounds, and the missing 12th compound will be dropped in all sample-source
comparisons if include_creosote = T. It is recommended that users first include creosote
to determine if it is an important source. If it is not important (determined by using TRUE),
then proceed using analyses that do not include creosote profiles.}
}
\value{
A data frame that contains
}
\description{
Conducts principle components analysis on the source and sample PAH profiles. This reduces the profiles
to manageable dimensions. All important PCA axes are then used to calculate the Euclidean distance between
source and samples to determine likely sources.
}
